Rewrote STATE_PLAYER_* states so that they properly receive an entire color sequence and set gameState appropriately.
TODO: There is currently no visual sign that the program received the player's input. Would be nice if the color blinked and made a sound (with each color having it's own sound) after it was clicked. Also a sidebar that displays the correct color sequence so far in a player's turn would be nice.